LayoutTests/platform/ios/imported/w3c/web-platform-tests/html/semantics/disabled-elements/disabled-event-dispatch.tentative-expected.txt

hello world child  hello world child

FAIL Testing mousedown events when clicking child of disabled button. assert_equals: parent element received mousedown events expected false but got true
FAIL Testing mousedown events when clicking child of disabled my-control. assert_equals: parent element received mousedown events expected false but got true
PASS Testing mousedown events when clicking disabled button.
PASS Testing mousedown events when clicking disabled my-control.
FAIL Testing mouseup events when clicking child of disabled button. assert_equals: parent element received mouseup events expected false but got true
FAIL Testing mouseup events when clicking child of disabled my-control. assert_equals: parent element received mouseup events expected false but got true
PASS Testing mouseup events when clicking disabled button.
PASS Testing mouseup events when clicking disabled my-control.
FAIL Testing pointerdown events when clicking child of disabled button. assert_equals: parent element received pointerdown events expected true but got false
FAIL Testing pointerdown events when clicking child of disabled my-control. assert_equals: parent element received pointerdown events expected true but got false
FAIL Testing pointerdown events when clicking disabled button. assert_equals: parent element received pointerdown events expected true but got false
FAIL Testing pointerdown events when clicking disabled my-control. assert_equals: parent element received pointerdown events expected true but got false
FAIL Testing pointerup events when clicking child of disabled button. assert_equals: parent element received pointerup events expected true but got false
FAIL Testing pointerup events when clicking child of disabled my-control. assert_equals: parent element received pointerup events expected true but got false
FAIL Testing pointerup events when clicking disabled button. assert_equals: parent element received pointerup events expected true but got false
FAIL Testing pointerup events when clicking disabled my-control. assert_equals: parent element received pointerup events expected true but got false
FAIL Testing click events when clicking child of disabled button. assert_equals: parent element received click events expected false but got true
FAIL Testing click events when clicking child of disabled my-control. assert_equals: parent element received click events expected false but got true
PASS Testing click events when clicking disabled button.
PASS Testing click events when clicking disabled my-control.
